score final

+++ Juego/Assets/ScoreScript.cs
@@ -3,8 +3,12 @@ using UnityEngine.UI;
 
 public class ScoreScript : MonoBehaviour
 {
-    public static int scoreValue = 0;
+    public static double scoreValue = 0;
     public Text scoreText;
+
+    //public Text puntText;
+    //double punt = scoreValue * GameControl.health;
+
     // Start is called before the first frame update
     void Start()
     {
